Move all frontend locale-related code to `locales/`
`index.ts` re-exports what can be consumed by the app and nothing else
