+++ app/javascript/packs/application.js
@@ -2,7 +2,7 @@ import './public-path';
 import main from "mastodon/main"
 
 import { start } from '../mastodon/common';
-import { loadLocale } from '../mastodon/load_locale';
+import { loadLocale } from '../mastodon/locales';
 import { loadPolyfills } from '../mastodon/polyfills';
 
 start();
