+++ app/javascript/packs/public.jsx
@@ -15,8 +15,7 @@ import { start } from '../mastodon/common';
 import { timeAgoString }  from '../mastodon/components/relative_timestamp';
 import emojify  from '../mastodon/features/emoji/emoji';
 import loadKeyboardExtensions from '../mastodon/load_keyboard_extensions';
-import { loadLocale } from '../mastodon/load_locale';
-import { getLocale }  from '../mastodon/locales';
+import { loadLocale, getLocale } from '../mastodon/locales';
 import { loadPolyfills } from '../mastodon/polyfills';
 import ready from '../mastodon/ready';
